backend/pyproject.toml
[tool.poetry]
name = "application"
version = "0.1.0"
description = "A simple app for managing loan defaulters"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Flask-SQLAlchemy = "^2.5.1"
Flask-Cors = "^3.0.10"
Flask-JWT-Extended = "^4.4.4"
flask-marshmallow = "^0.14.0"
Flask-Migrate = "^3.1.0"
python-dotenv = "^0.20.0"
psycopg2 = "^2.9.3"

[tool.poetry.dev-dependencies]
black = {version = "^22.6.0", allow-prereleases = true}
isort = "^5.10.1"
mypy = "^0.971"
vulture = "^2.5"
pytest = "^7.1.2"
flake8 = "^5.0.4"
flake8-print = "^5.0.0"
pep8-naming = "^0.13.2"

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.isort]
combine_as_imports = true
force_grid_wrap = 0
include_trailing_comma = true
known_first_party = "backend"
line_length = 100
multi_line_output = 3
no_lines_before = "LOCALFOLDER"
order_by_type = false

[tool.mypy]
allow_redefinition = true
disallow_untyped_defs = true
warn_unused_ignores = true

[tool.vulture]
min_confidence = 100



[tool.poetry.scripts]
format = "scripts:format"
reformat = "scripts:reformat"
start = "wsgi:app.run"
lint = "scripts:lint"



[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
